+++ android/src/main/kotlin/com/byteowls/capacitor/filesharer/FileSharerProvider.kt
@@ -1,0 +1,5 @@
+package com.byteowls.capacitor.filesharer
+
+import androidx.core.content.FileProvider
+
+class FileSharerProvider : FileProvider()